continued work on post-image to try and work around the 4k body limitation. it almost works. this was supposed to be easy.

+++ post-image.py
@@ -11,31 +11,51 @@ import modem as modem_module
 modem = modem_module.modem()
 modem.init()
 
-image_file="/dev/shm/image-{}".format(datetime.datetime.isoformat(datetime.datetime.now()))
+if len(sys.argv) > 1 and sys.argv[1] == "old":
+    print("Debug mode: reading /dev/shm/old image")
+    image_file = "/dev/shm/old"
+else:
+    image_file="/dev/shm/image-{}".format(datetime.datetime.isoformat(datetime.datetime.now()))
+    subprocess.run(["raspistill -t 2000 -o {}.png -e png".format(image_file)], shell=True, check=True)
+    subprocess.run(["cwebp -q 50 {}.png -o {}.webp".format(image_file,image_file)], shell=True, check=True)
+    os.unlink("{}.png".format(image_file))
 
-subprocess.run(["raspistill -t 2000 -o {}.png -e png".format(image_file)], shell=True, check=True)
-subprocess.run(["cwebp -q 50 {}.png -o {}.webp".format(image_file,image_file)], shell=True, check=True)
 image = None
 with open("{}.webp".format(image_file), "rb") as in_file:
     image = in_file.read()
-os.unlink("{}.png".format(image_file))
 #sys.exit(11)
 
-while True:
-    modem.write_noblock('AT')
-    rcv = modem.port.read(100)
-    if len(rcv) > 0:
-        print(rcv)
-        break
-    print(".", end='', flush=True)
-    time.sleep(1)
-
-
-modem.lte_configure()
-modem.lte_connect()
-modem.print_status()
-
-url = "http://hacks.v9n.us/sim800c/?image=1"
+url_template = "http://hacks.v9n.us/sim800c/?image={}&id={}"
 host = "http://hacks.v9n.us"
-modem.lte_http_post(host, url, image)
-modem.lte_disconnect()
+
+def chunks(lst, n):
+    """ Yield successive n-sized chunks from lst.
+        https://stackoverflow.com/a/312464 """
+    for i in range(0, len(lst), n):
+        yield lst[i:i + n]
+
+iter = 0
+print("DBG: len webp={}".format(len(image)))
+for chunk in chunks(image, 3500):
+    print("DBG: {}, len={}".format(iter, len(chunk)))
+    print(chunk[:50])
+    print()
+    iter += 1
+
+#sys.exit(42)
+
+id = 0
+iter = 0
+for chunk in chunks(image, 1000):
+    modem.lte_configure()
+    modem.lte_connect()
+    modem.print_status()
+    print("*****†**********†chunk iter {} id {}".format(iter, id))
+    url = url_template.format('new' if id == 0 else 'append', id)
+    response = modem.lte_http_post(host, url, chunk)
+    part1 = response[response.find('id='):]
+    part2 = part1[:part1.find("\r\n")]
+    id = part2.split("=")[1].strip()
+    iter += 1
+    #time.sleep(30)
+    modem.lte_disconnect()
